商品接口提交

+++ trade-center/trade-center-servie/src/main/java/com/ai/bdex/dataexchange/tradecenter/service/interfaces/gds/IGdsInfoSV.java
@@ -7,5 +7,5 @@ import com.ai.bdex.dataexchange.tradecenter.dao.model.GdsInfo;
  */
 public interface IGdsInfoSV {
 
-    public GdsInfo queryGdsInfoById(Long gdsId) throws Exception;
+    public GdsInfo queryGdsInfoById(Integer gdsId) throws Exception;
 }
